refactor: move /ws/portal calls into common service modules for shop and events
