+++ app/[tenant]/[workspace]/(subapps)/events/common/service/index.ts
@@ -1,15 +1,18 @@
 import axios from 'axios';
-import type {Cloned} from '@/types/util';
 
 // ---- CORE IMPORTS ---- //
 import {t} from '@/locale/server';
 import type {TenantConfig} from '@/tenant';
 import {ID} from '@/types';
 import {PortalWorkspace} from '@/orm/workspace';
+import {getSession} from '@/auth';
+import {findWorkspace} from '@/orm/workspace';
+import {ActionResponse} from '@/types/action';
+import type {Client} from '@/goovee/.generated/client';
+import type {Cloned} from '@/types/util';
 
 // ---- LOCAL IMPORTS ---- //
 import {error} from '@/subapps/events/common/utils';
-import {ActionResponse} from '@/types/action';
 
 export async function createInvoice({
   workspace,
@@ -69,5 +72,72 @@ export async function createInvoice({
         'An error occurred while processing your invoice. Please try again later.',
       ),
     );
+  }
+}
+
+export async function findProductsFromWS({
+  workspaceURL,
+  client,
+  config,
+  eventId,
+}: {
+  workspaceURL: string;
+  eventId: string;
+  client: Client;
+  config: TenantConfig;
+}) {
+  if (!workspaceURL && eventId) {
+    return null;
+  }
+
+  if (!config?.aos?.url) {
+    return [];
+  }
+
+  const session = await getSession();
+  const user = session?.user;
+
+  const workspace = await findWorkspace({
+    url: workspaceURL,
+    user,
+    client,
+  });
+
+  if (!workspace) {
+    return {
+      error: true,
+      message: await t('Invalid workspace'),
+    };
+  }
+  const {aos} = config;
+
+  const ws = `${aos.url}/ws/portal/event/price`;
+
+  const partnerId = user?.id;
+
+  try {
+    const reqBody = {
+      eventId,
+      partnerWorkspaceId: workspace.id,
+      partnerId,
+    };
+    const res = await axios
+      .post(ws, reqBody, {
+        auth: {
+          username: aos.auth.username,
+          password: aos.auth.password,
+        },
+      })
+      .then(({data}) => data);
+
+    if (res?.data?.status === -1) {
+      console.log('Error:', res);
+      return null;
+    }
+
+    return res?.data || null;
+  } catch (err) {
+    console.log('Error:', err);
+    return null;
   }
 }
