+++ app/[tenant]/[workspace]/(subapps)/shop/common/service/index.ts
@@ -1,0 +1,221 @@
+import axios from 'axios';
+import {headers} from 'next/headers';
+
+// ---- CORE IMPORTS ---- //
+import {t} from '@/locale/server';
+import {PortalWorkspace} from '@/orm/workspace';
+import {Cloned} from '@/types/util';
+import type {Tenant} from '@/tenant';
+import type {Client} from '@/goovee/.generated/client';
+import {computeTotal} from '@/utils/cart';
+import {TENANT_HEADER} from '@/proxy';
+import {getSession} from '@/auth';
+import {manager} from '@/tenant';
+import {MAIN_PRICE} from '@/constants';
+
+// ---- LOCAL IMPORTS ---- //
+import {findProduct} from '@/subapps/shop/common/orm/product';
+import {formatNumber} from '@/subapps/shop/common/utils/order';
+import {calculateAdvanceAmount} from '@/utils/payment';
+
+export async function createOrder({
+  cart,
+  workspace,
+  user,
+  client,
+  config,
+  paymentModeId,
+}: {
+  cart: any;
+  workspace: PortalWorkspace | Cloned<PortalWorkspace>;
+  user: NonNullable<any>;
+  client: Client;
+  config: Tenant['config'];
+  paymentModeId?: string;
+}) {
+  const {aos} = config;
+  const ws = `${aos.url}/ws/portal/orders/order`;
+
+  const computedProducts = await Promise.all(
+    cart.items.map((i: any) =>
+      findProduct({id: i.product, workspace, user, client}),
+    ),
+  );
+
+  const $cart = {
+    ...cart,
+    items: cart.items.map((i: any) => ({
+      ...i,
+      computedProduct: computedProducts.find(
+        cp => Number(cp?.product?.id) === Number(i.product),
+      ),
+    })),
+  };
+
+  const {total} = computeTotal({cart: $cart, workspace, formatNumber});
+
+  const {id, isContact, mainPartnerId} = user;
+  const partnerId = isContact && mainPartnerId ? mainPartnerId : id;
+  const contactId = isContact && mainPartnerId ? id : undefined;
+
+  const {invoicingAddress, deliveryAddress} = cart;
+  const payInAdvance = workspace.config?.payInAdvance;
+  const advancePaymentPercentage = workspace.config?.advancePaymentPercentage;
+
+  const paidAmount =
+    payInAdvance && Number(advancePaymentPercentage) > 0
+      ? calculateAdvanceAmount({
+          amount: Number(total),
+          percentage: Number(advancePaymentPercentage),
+          payInAdvance,
+        }).toString()
+      : Number(total).toString();
+
+  const isAtiPricing = workspace?.config?.mainPrice === MAIN_PRICE.ATI;
+
+  const payload = {
+    partnerId,
+    contactId,
+    shipping: 0,
+    total,
+    inAti: isAtiPricing,
+    items: $cart.items.map((i: any) => {
+      const {computedProduct, note, quantity} = i;
+      if (!computedProduct) return null;
+      const {product, price} = computedProduct;
+      return {
+        productId: product?.id,
+        note: note || '',
+        quantity,
+        price: isAtiPricing ? price?.ati : price?.wt,
+      };
+    }),
+    workspaceId: workspace.id,
+    invocingPartnerAddressId: invoicingAddress,
+    deliveryPartnerAddressId: deliveryAddress,
+    paidAmount,
+    paymentModeId,
+  };
+
+  const res = await axios.post(ws, payload, {
+    auth: {username: aos.auth.username, password: aos.auth.password},
+  });
+
+  if (res?.data?.status === -1) {
+    throw new Error(
+      res?.data?.message
+        ? await t(res.data.message)
+        : await t('Order creation failed. Please try again.'),
+    );
+  }
+
+  return res?.data;
+}
+
+export async function requestOrder({
+  cart,
+  workspace,
+  type = 'order',
+}: {
+  cart: any;
+  workspace: PortalWorkspace | Cloned<PortalWorkspace>;
+  type?: 'quotation' | 'order';
+}) {
+  const tenantId = (await headers()).get(TENANT_HEADER);
+
+  if (!tenantId) {
+    return null;
+  }
+
+  if (!cart?.items?.length) return null;
+
+  const tenant = await manager.getTenant(tenantId);
+
+  if (!tenant?.config?.aos?.url) return null;
+
+  const {aos} = tenant.config;
+  const {client} = tenant;
+
+  const ws = `${aos.url}/ws/portal/orders/${type}`;
+
+  const session = await getSession();
+  const user = session?.user;
+
+  if (!(session && workspace && workspace.config)) return null;
+
+  try {
+    const computedProducts = (
+      await Promise.all(
+        cart.items.map((i: any) =>
+          findProduct({id: i.product, workspace, user, client}),
+        ),
+      )
+    ).filter(Boolean);
+
+    const $cart = {
+      ...cart,
+      items: [
+        ...cart?.items?.map((i: any) => ({
+          ...i,
+          computedProduct: computedProducts.find(
+            cp => Number(cp?.product?.id) === Number(i.product),
+          ),
+        })),
+      ],
+    };
+
+    const {total} = computeTotal({cart: $cart, workspace, formatNumber});
+
+    let partnerId, contactId;
+
+    if (user) {
+      const {id, isContact, mainPartnerId} = user;
+      if (isContact && mainPartnerId) {
+        partnerId = mainPartnerId;
+        contactId = id;
+      } else {
+        partnerId = id;
+      }
+    }
+    const {invoicingAddress, deliveryAddress} = cart;
+    const isAtiPricing = workspace?.config?.mainPrice === MAIN_PRICE.ATI;
+
+    const payload = {
+      partnerId,
+      contactId,
+      shipping: 0,
+      total,
+      inAti: isAtiPricing,
+      items: $cart.items.map((i: any) => {
+        const {computedProduct, note, quantity} = i;
+        if (!computedProduct) return null;
+        const {product, price} = computedProduct;
+        return {
+          productId: product?.id,
+          note: note || '',
+          quantity,
+          price: isAtiPricing ? price?.ati : price?.wt,
+        };
+      }),
+      workspaceId: workspace.id,
+      invocingPartnerAddressId: invoicingAddress,
+      deliveryPartnerAddressId: deliveryAddress,
+    };
+
+    const res = await axios.post(ws, payload, {
+      auth: {
+        username: aos.auth.username,
+        password: aos.auth.password,
+      },
+    });
+
+    if (res?.data?.status === -1) {
+      return null;
+    }
+
+    return res?.data;
+  } catch (err) {
+    console.error(err);
+    return null;
+  }
+}
